+++ love/src/Utils.js
@@ -1176,7 +1176,7 @@ export default class ManagerInterface {
     });
   }
 
-  static updateScriptSchema(id, configSchema) {
+  static updateScriptSchema(id, configSchema, schema) {
     const token = ManagerInterface.getToken();
     if (token === null) {
       return new Promise((resolve) => resolve(false));
@@ -1187,6 +1187,7 @@ export default class ManagerInterface {
       headers: this.getHeaders(),
       body: JSON.stringify({
         config_schema: configSchema,
+        schema: schema,
       }),
     }).then((response) => {
       if (response.status >= 500) {
@@ -1198,7 +1199,7 @@ export default class ManagerInterface {
       }
       if (response.status === 400) {
         return response.json().then((resp) => {
-          toast.error(resp.ack);
+          toast.error(resp.title);
           return false;
         });
       }
@@ -1208,7 +1209,7 @@ export default class ManagerInterface {
     });
   }
 
-  static postScriptConfiguration(scriptPath, scriptType, configName, configSchema) {
+  static postScriptConfiguration(scriptPath, scriptType, configName, configSchema, schema) {
     const token = ManagerInterface.getToken();
     if (token === null) {
       return new Promise((resolve) => resolve(false));
@@ -1222,6 +1223,7 @@ export default class ManagerInterface {
         script_type: scriptType,
         config_name: configName,
         config_schema: configSchema,
+        schema: schema,
       }),
     }).then((response) => {
       if (response.status >= 500) {
@@ -1233,7 +1235,7 @@ export default class ManagerInterface {
       }
       if (response.status === 400) {
         return response.json().then((resp) => {
-          toast.error(resp.ack);
+          toast.error(resp.title);
           return false;
         });
       }
@@ -1825,4 +1827,16 @@ export function parseToSALFormat(data) {
 export function copyToClipboard(text, effect) {
   navigator.clipboard.writeText(text);
   if (effect) effect();
+}
+
+/**
+ * Function to trim a string to a specified length
+ * @param {string} string text to be trimmed
+ * @param {number} length length to be trimmed, 100 by default
+ */
+export function trimString(string, length = 100) {
+  if (string.length > length) {
+    return `${string.substring(0, length)}...`;
+  }
+  return string;
 }